Modified: -displaying dosage checkbox depends on countable param in medicine type -added condition stopping deleting medicine count when doesn't exist -adjust spinner displays dialog window -some string values

+++ app/src/main/java/smartcity/kni/wirtualnaapteczka/net/database/SQLiteDatabaseHelper.java
@@ -133,7 +133,8 @@ public class SQLiteDatabaseHelper {
         for(Information i: medicine.getInformationList()) {
             daoSession.getInformationDao().delete(i);
         }
-        daoSession.getMedicine_CountDao().delete(medicine.getMedicine_Count());
+        if(medicine.getMedicine_Count() != null)
+            daoSession.getMedicine_CountDao().delete(medicine.getMedicine_Count());
         ////////////////////////
 
         daoSession.getMedicineDao().deleteByKey(idMedicine);
